requirements.txt
selenium==3.141.0
Appium-Python-Client==1.0.2
logbook==1.5.3
configparser==5.0.1
ddt==1.4.1